Added activity to write the card name. A little bit clean up.
