+++ .idea/misc.xml
@@ -7,7 +7,7 @@
         <entry key="..\:/MyFinances/app/src/main/res/layout/activity_log_in2.xml" value="0.22760416666666666" />
         <entry key="..\:/MyFinances/app/src/main/res/layout/activity_main.xml" value="0.21354166666666666" />
         <entry key="..\:/MyFinances/app/src/main/res/layout/activity_register.xml" value="0.25" />
-        <entry key="..\:/MyFinances/app/src/main/res/layout/activity_save_card.xml" value="0.264" />
+        <entry key="..\:/MyFinances/app/src/main/res/layout/activity_save_card.xml" value="0.4" />
       </map>
     </option>
   </component>
