+++ app/src/main/java/pl/edu/pja/myfinances/MainActivity.kt
@@ -3,46 +3,13 @@ package pl.edu.pja.myfinances
 import android.content.Intent
 import android.os.Bundle
 import android.view.View
-import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import com.google.firebase.auth.FirebaseAuth
-import com.google.firebase.database.FirebaseDatabase
-import com.journeyapps.barcodescanner.ScanContract
-import com.journeyapps.barcodescanner.ScanIntentResult
-import com.journeyapps.barcodescanner.ScanOptions
 import pl.edu.pja.myfinances.databinding.ActivityMainBinding
-import pl.edu.pja.myfinances.model.Card
-
 
 class MainActivity : AppCompatActivity() {
     private val binding by lazy { ActivityMainBinding.inflate(layoutInflater) }
     private lateinit var auth: FirebaseAuth
-    private val barcodeLauncher = registerForActivityResult(ScanContract()) { result: ScanIntentResult ->
-        if (result.contents == null) {
-            Toast.makeText(
-                this,
-                "Cancelled",
-                Toast.LENGTH_LONG
-            ).show()
-        } else {
-            Toast.makeText(
-                this,
-                "Scanned: " + result.contents,
-                Toast.LENGTH_LONG
-            ).show()
-            startActivity(Intent(this, LogInActivity::class.java))
-            FirebaseDatabase.getInstance()
-                .getReference(auth.uid!!)
-                .child("cards")
-                .push()
-                .setValue(
-                    Card(
-                        "KARTA",
-                        result.contents
-                    )
-                )
-        }
-    }
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -56,14 +23,7 @@ class MainActivity : AppCompatActivity() {
     }
 
     fun scanBarCode(view: View) {
-        val options = ScanOptions()
-//        options.setDesiredBarcodeFormats(ScanOptions.ONE_D_CODE_TYPES)
-//        options.setPrompt("Scan a barcode")
-//        options.setCameraId(0) // Use a specific camera of the device
-//
-//        options.setBeepEnabled(false)
-//        options.setBarcodeImageEnabled(true)
-        val aaa = barcodeLauncher.launch(options)
+        startActivity(Intent(this, SaveCardActivity::class.java))
     }
 
     fun signOut(view: View) {
